Games for tournaments

+++ src/main/kotlin/UNQ/TTIP/GOAT/dao/GameDAO.kt
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface GameDAO : JpaRepository<Game, Int?> {
+interface GameDAO : JpaRepository<Game, Long?> {
 
     fun findByTournamentId(tournament_id: Long): MutableList<Game>
     fun findByTeamsTeamIdContaining(teams_team_id: Long): MutableList<Game>
